Вынесение модалки в отдельную компоненту
