+++ pages/Main/Main.js
@@ -19,6 +19,7 @@ import { CandleChart } from "react-native-wagmi-charts"
 import { FetchCoinHistoricalData, GetMarketData } from "../../components/Api/Api"
 
 import Ionicons from "react-native-vector-icons/Ionicons"
+import { ModalView } from "../../components/ModalView/ModalView"
 const Main = () => {
   const [search, setSearch] = useState("")
   const [refreshing, setRefreshing] = useState(false)
@@ -31,22 +32,8 @@ const Main = () => {
   const [errorMessage, setErrorMessage] = useState(null)
   const switchChartDays = useSelector(daysSelector)
 
-  // const [dropdownVisible, setDropdownVisible] = useState(false)
-  // const toggleDropdown = () => {
-  //   setDropdownVisible(!dropdownVisible)
-  // }
   const [modal, setModal] = useState(false)
-
   const toggleModal = () => {
-    setModal(!modal)
-  }
-
-  const handlePress = (e) => {
-    if (e === 1) {
-      console.log("open small")
-    } else {
-      console.log("open large")
-    }
     setModal(!modal)
   }
 
@@ -149,26 +136,8 @@ const Main = () => {
 
         {/* logo-codepen, logo-buffer, tv, pulse, menu, list, analytics , grid, */}
       </View>
-      <Modal
-        transparent={true}
-        visible={modal}
-        animationType='fade'
-        onRequestClose={handlePress}
-      >
-        <View style={styles.dropdown}>
-          <View style={styles.dropbox}>
-            <TouchableOpacity style={styles.dropdownItem} onPress={() => handlePress(1)}>
-              <Text style={styles.dropdownText}>Compact</Text>
-              <Ionicons style={styles.changePoint} name='grid' size={20} />
-            </TouchableOpacity>
-          </View>
 
-          <TouchableOpacity style={styles.dropdownItem} onPress={() => handlePress(2)}>
-            <Text style={styles.dropdownText}>Medium</Text>
-            <Ionicons style={styles.changePoint} name='list' size={20} />
-          </TouchableOpacity>
-        </View>
-      </Modal>
+      <ModalView modal={modal} setModal={setModal} />
 
       {flagForLoader ? (
         <ActivityIndicator size='large' color='red' />
